add DatabaseUtils class to split /create-order into functions
